Add new tutorial link
This is a tutorial I wrote that shows how to use the storybook-dark-mode addon to enable a dark mode toggle option for Storybook, and then how to connect this dark mode toggle to the React components being rendered in preview. So when Storybook is toggled to dark mode, the component being rendered in preview changes as well. The tutorial covers everything involved, from installing Storybook to configuring the React components to switch between dark and light mode. Any edits to the description or title I added in the README are welcome. And of course if this doesn’t fit in with the theme of what you’re going for here feel free to ignore — thanks!